% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/caterpillar_elo.R
\name{caterpillar_elo}
\alias{caterpillar_elo}
\title{caterpillar plots for elo model}
\usage{
caterpillar_elo(res, id = NULL, include_k = TRUE, mfrow = c(2, 2))
}
\arguments{
\item{res}{object from \code{\link{elo_seq_bayes}}}

\item{id}{character vector with ids to include (default is \code{NULL}:
plot a random sample of all individuals, as constrained by
\code{mfrow})}

\item{include_k}{plot k parameter first (default is \code{TRUE})}

\item{mfrow}{layout of plot (default is \code{c(2, 2)})}
}
\value{
a plot
}
\description{
caterpillar plots for elo model
}
\examples{
x <- generate_interactions(n_ind = 6, n_int = c(100), use_letters = TRUE)
x <- data.frame(x)
standat <- prep_seq(winner = x$winner, loser = x$loser, Date = x$date)
res <- elo_seq_bayes(standat = standat, refresh = 0, parallel_chains = 2)
par(mar = c(2, 2, 1, 1), las = 1)
# k + plus 3 individuals
caterpillar_elo(res)
# only k
caterpillar_elo(res, mfrow = c(1, 1))
# two specific individuals
caterpillar_elo(res, id = c(x$winner[1], x$loser[1]), include_k = FALSE, mfrow = c(1, 2))
}
