% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/elo_seq_bayes.R
\name{elo_seq_bayes}
\alias{elo_seq_bayes}
\title{Elo-rating with Bayesian flavor}
\usage{
elo_seq_bayes(standat, quiet = FALSE, ...)
}
\arguments{
\item{standat}{output of \code{\link{prep_seq}}}

\item{quiet}{logical, default is \code{FALSE}. If set to \code{TRUE}:
try to swallow all output cmdstanr prints
to the console (which can be annoying in some contexts).}

\item{...}{additional arguments to \code{cmdstanr}'s
\code{\link[cmdstanr]{sample}} method (typically
\code{parallel_chains}, \code{refresh},
\code{iter_warmup}, \code{iter_sampling},
\code{seed}, \code{adapt_delta},
and/or \code{step_size})}
}
\value{
a list with two items, where the first (\code{$mod_res}) is the
the \code{cmdstanr} model object (see
\code{\link[cmdstanr]{CmdStanModel}}). The second list item
(\code{$standat}) is the list supplied as stan data (which
is required for further processing).
}
\description{
Elo-rating with Bayesian flavor
}
\examples{
x <- EloRating::randomsequence(nID = 7, avgIA = 30)$seqdat
standat <- prep_seq(winner = x$winner, loser = x$loser, Date = x$Date)
res <- elo_seq_bayes(standat = standat, refresh = 0, parallel_chains = 2)
summary(res)

# incorporating presence and ties
x <- EloRating::randomsequence(nID = 8, avgIA = 20,
                               presence = c(0.2, 0.4), ties = 0.3)
winner <- x$seqdat$winner
loser <- x$seqdat$loser
Date <- x$seqdat$Date
presence <- x$pres
draws <- x$seqdat$Draw
standat <- prep_seq(winner = winner, loser = loser, Date = Date,
                    presence = presence, draws = draws,
                    estimate_startspread = TRUE)
res <- elo_seq_bayes(standat = standat, refresh = 0, chains = 3,
                     iter_sampling = 333)
summary(res)

x <- EloRating::randomsequence(nID = 8, avgIA = 20)$seqdat
x$intensity <- sample(c("contact", "mild"), nrow(x), TRUE)
standat <- prep_seq(winner = x$winner, loser = x$loser, Date = x$Date,
                    intensity = x$intensity)
res <- elo_seq_bayes(standat = standat, refresh = 0, parallel_chains = 2)
summary(res)
}
\references{
The underlying is Stan code is heavily based on this paper:

\insertRef{goffe2018}{EloRating.Bayes}
}
