% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_extract_index.R
\name{find_extract_index}
\alias{find_extract_index}
\alias{find_extract_index2}
\title{index position for extraction dates}
\usage{
find_extract_index(idates, edates = NULL)
}
\arguments{
\item{idates}{character or Date vector with interaction
dates (of the observed interactions)}

\item{edates}{character or Date vector of dates for which corresponding
ratings are to be extracted. Default is \code{NULL}, i.e.,
extract ratings after first interaction and ratings after
last interaction date.}

\item{standat}{stan data list}

\item{id}{vector with ids}
}
\value{
a named integer vector
}
\description{
index position for extraction dates
}
\examples{
idates <- c("2000-01-04", "2000-03-27", "2000-04-01", "2000-04-01", "2000-04-24")
find_extract_index(idates)
# return index of first date (with message)
find_extract_index(idates, edates = "1999-01-01")
find_extract_index(idates, edates = c("2000-01-09", "2000-04-15"))
find_extract_index(idates, edates = c("2000-04-09", "2020-04-15"))

data(adv, package = "EloRating")
standat <- prep_seq(winner = adv$winner, loser = adv$loser, Date = adv$Date)
find_extract_index2(standat, c("a", "b"))
}
