% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/date_helper.R
\name{find_origin}
\alias{find_origin}
\title{find origin for as.Date}
\usage{
find_origin(xdate)
}
\arguments{
\item{xdate}{a \code{Date}}
}
\value{
character
}
\description{
find origin for as.Date
}
\examples{
find_origin(as.Date("2000-10-10"))
}
