% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_interactions.R
\name{generate_interactions}
\alias{generate_interactions}
\title{generate dominance interaction sequence with varying steepness}
\usage{
generate_interactions(
  n_ind = 10,
  n_int = c(100, 100),
  steep = c(0.9, 0.2),
  shuffle_dates = FALSE
)
}
\arguments{
\item{n_ind}{integer, number of individuals (default is 10)}

\item{n_int}{integer vector, number of interactions per steepness period}

\item{steep}{numeric vector, steepness per period (needs to be the same
length as \code{n_int})}

\item{shuffle_dates}{logical, create regular date sequence or shuffle dates,
such that some dates are repeated and others don't
occur}
}
\value{
a data frame with interaction sequence
}
\description{
steepness changes are induced via periods determined by interaction numbers
}
\examples{
# 15 interactions with steepness 0.9
# followed by 10 interactions with steepness 0.3
generate_interactions(n_ind = 6, n_int = c(15, 10), steep = c(0.9, 0.3))

}
