% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/group_mixture.R
\name{group_mixture}
\alias{group_mixture}
\title{generate interaction data with migrating individuals}
\usage{
group_mixture(
  n_ind = 10,
  n_mig = 2,
  n_grp = 2,
  n_abs = 2,
  n_int = 32,
  mig_mode = c("average", "inherit", "bottom", "top"),
  stop_at_iter = 500,
  silent = FALSE,
  fail_at_stop = TRUE
)
}
\arguments{
\item{n_ind}{integer, number of individuals}

\item{n_mig}{integer, number of migrating individuals (must be
smaller than \code{n_int})}

\item{n_grp}{integer, number of groups}

\item{n_abs}{integer, number of individuals that were at some point
absent from the population (max absence time
is half the study duration)}

\item{n_int}{integer, number of interactions}

\item{mig_mode}{determining the mode of how the migrators integrate
in their new group}

\item{stop_at_iter}{numeric, number of iterations until function fails}

\item{silent}{logical, omit info messages or not}

\item{fail_at_stop}{logical, if function is unsuccessful: return error or \code{NULL}}
}
\value{
a list, where the first item is a data.frame and the second one is
a presence/group allocation matrix
}
\description{
each individual can migrate multiple times (and also back and forth)
}
\details{
The 'inherit' mode for migration is a bit misleading at the moment.
In fact, the start ratings are inherited, i.e. if an individual rises from
a low start rating to a high rating within its group and then migrates,
it 'inherits' the low rating.
Within the simulated data sets this is okay, because ratings are stable
within groups, but in natural settings this might not be realistic...
}
\examples{
group_mixture()
}
