% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/last_interaction.R
\name{last_interaction}
\alias{last_interaction}
\title{find last interaction date of an individual}
\usage{
last_interaction(test_ids, test_dates, interaction_dates, winner, loser)
}
\arguments{
\item{test_ids}{character vector with id codes}

\item{test_dates}{character or Date vector: dates to be tested (must be
same length as \code{test_ids})}

\item{interaction_dates}{character or Date vector of observed interaction}

\item{winner, loser}{character vector with id codes of winners and losers}
}
\value{
a character vector with dates
}
\description{
find last interaction date of an individual
}
\examples{
data(adv, package = "EloRating")
# g's first interactions:
adv$Date[adv$winner == "g" | adv$loser == "g"][1:4]
last_interaction("g", "2010-01-07", interaction_dates = adv$Date, winner = adv$winner, loser = adv$loser)
last_interaction("g", "2010-01-08", interaction_dates = adv$Date, winner = adv$winner, loser = adv$loser)
last_interaction("g", "2010-01-01", interaction_dates = adv$Date, winner = adv$winner, loser = adv$loser)
}
