% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_stan_foos.R
\name{list_stan_foos}
\alias{list_stan_foos}
\alias{cat_foo}
\title{list and print (via \code{cat()}) Stan functions in the package}
\usage{
list_stan_foos()

cat_foo(foo)
}
\arguments{
\item{foo}{name of stan function}
}
\value{
textual output
}
\description{
list and print (via \code{cat()}) Stan functions in the package
}
