% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/longitudinal_steepness.R
\name{longitudinal_steepness}
\alias{longitudinal_steepness}
\title{steepness at specific dates}
\usage{
longitudinal_steepness(res, dates)
}
\arguments{
\item{res}{result of \code{\link{elo_seq_bayes}}}

\item{dates}{character of target date(s) (YYYY-MM-DD)}
}
\value{
a vector with one column for each entry in \code{dates}
}
\description{
steepness at specific dates
}
\examples{
# 3 periods with varying steepness:
x <- generate_interactions(n_ind = 12, n_int = c(200, 200, 200), steep = c(0.9, 0.2, 0.6))
res <- elo_seq_bayes(prep_seq(winner = x$winner,
                              loser = x$loser,
                              Date = x$date),
                     parallel_chains = 4)
dates <- c(x$date[max(which(x$set == 1))],
           x$date[max(which(x$set == 2))],
           x$date[max(which(x$set == 3))])
r <- longitudinal_steepness(res = res, dates = dates)
colMeans(r)
}
