% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principled_interactions.R
\name{plot_principled}
\alias{plot_principled}
\title{plotting ratings from interaction simulations}
\usage{
plot_principled(rating_grid, idata = NULL)
}
\arguments{
\item{rating_grid}{results from \code{\link{make_rating_grid}}}

\item{idata}{results from \code{\link{principled_interactions}}}
}
\value{
a plot
}
\description{
plotting ratings from interaction simulations
}
\examples{
rating_grid <- make_rating_grid(n_ind = 7, n_days = 100, drifts = 3)
plot_principled(rating_grid)

idata <- principled_interactions(rating_grid = rating_grid, n_int = 27)
plot_principled(rating_grid, idata)

rating_grid <- make_rating_grid(n_ind = 7, n_days = 100, start_sd = 0, drifts = 3)
plot_principled(rating_grid)
}
