% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores.R
\name{plot_scores}
\alias{plot_scores}
\title{cross-sectional plot of rating posteriors}
\usage{
plot_scores(res, targetdate = NULL, color = TRUE)
}
\arguments{
\item{res}{result of \code{\link{elo_seq_bayes}}}

\item{targetdate}{character of target date (YYYY-MM-DD):
must be of length = 1}

\item{color}{logical, default is \code{TRUE} where individuals
get color-coded. If \code{FALSE}: a gray scale is used.
It is also possible to hand over a vector with colors,
which then must correspond in length to the number of individuals.}
}
\value{
a plot
}
\description{
cross-sectional plot of rating posteriors
}
\examples{
library(EloRating)
x <- randomsequence(nID = 7, avgIA = 100)$seqdat
standat <- prep_seq(winner = x$winner, loser = x$loser, Date = x$Date)
res <- elo_seq_bayes(standat = standat, refresh = 0, parallel_chains = 2)
plot_scores(res, targetdate = "2000-01-01")
title(main = "start ratings")
plot_scores(res, targetdate = rev(names(standat$idates))[1])
title(main = "ratings at sequence end")
}
