% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_scores_longitudinal.R
\name{plot_scores_longitudinal}
\alias{plot_scores_longitudinal}
\title{plot ratings longitudinally (experimental!)}
\usage{
plot_scores_longitudinal(
  res,
  date_range = NULL,
  unc_width = 0.89,
  resol = 11,
  plot_presence = NULL
)
}
\arguments{
\item{res}{result of \code{\link{elo_seq_bayes}}}

\item{date_range}{character of date range (YYYY-MM-DD): must be of length = 2;
if \code{NULL}: the entire detected range of dates is used.}

\item{unc_width}{numeric between 0 and 1, width of uncertainty around median
(default is 0.89).}

\item{resol}{numeric, number of steps}

\item{plot_presence}{optional presence matrix used for plotting (see \code{\link{sim_group_split}})}
}
\value{
a plot
}
\description{
plot ratings longitudinally (experimental!)
}
\examples{
x <- EloRating::randomsequence(nID = 7, avgIA = 50, presence = c(0.4, 0.6))
standat <- prep_seq(winner = x$seqdat$winner, loser = x$seqdat$loser,
                    Date = x$seqdat$Date, presence = x$pres)
res <- elo_seq_bayes(standat = standat, refresh = 0, parallel_chains = 4)
plot_scores_longitudinal(res, resol = 7)
}
