% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_steepness_longitudinal.R
\name{plot_steepness_longitudinal}
\alias{plot_steepness_longitudinal}
\title{plot steepness posteriors at multiple time steps throughout a sequence}
\usage{
plot_steepness_longitudinal(res, dates, vertical_stretch = 200)
}
\arguments{
\item{res}{result of \code{\link{elo_seq_bayes}}}

\item{dates}{character of target date(s) (YYYY-MM-DD)}

\item{vertical_stretch}{numeric, expansion of the posteriors along the
horizontal axis. Setting this parameter is just
for appearance and for now requires a bit of
trial and error. Its value will depend on how
many steepness posteriors are going to be displayed
and how long the date range is in the data.
The default is 200.}
}
\value{
a plot (and an invisible \code{NULL})
}
\description{
plot steepness posteriors at multiple time steps throughout a sequence
}
\examples{
idata <- generate_interactions(n_ind = 8, n_int = c(300, 300), steep = c(0.9, 0.2))
res <- elo_seq_bayes(prep_seq(winner = idata$winner, loser = idata$loser, Date = idata$date), parallel_chains = 4)
dates <- c(idata$date[max(which(idata$set == 1))], idata$date[max(which(idata$set == 2))])
plot_steepness_longitudinal(res = res, dates = dates)
}
