% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pp_check.R
\name{pp_check}
\alias{pp_check}
\title{posterior predictive checks}
\usage{
pp_check(res, n_samples = 50)
}
\arguments{
\item{res}{result from \code{\link{elo_seq_bayes}}}

\item{n_samples}{number of simulated draws to plot}
}
\value{
a plot
}
\description{
by individual
}
\details{
Red is the obseverd winning proportion of each individual.

Summaries of simulated draws are added in black (cross = median, line = 80\% range)
}
\examples{
res <- toydata()
pp_check(res, n_samples = 10)
}
