% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_new_outcome.R
\name{predict_new_outcome}
\alias{predict_new_outcome}
\title{predict outcome}
\usage{
predict_new_outcome(model, i1, i2, threshold = 0.8)
}
\arguments{
\item{model}{result from \code{\link{elo_seq_bayes}}}

\item{i1, i2}{two individuals for which to get the outcome}

\item{threshold}{default is 0.6: how certain are we about the outcome reaching this level of confidence}
}
\value{
a data frame and a plot
}
\description{
predict outcome
}
\examples{
data(adv, package = "EloRating")
standat <- prep_seq(winner = adv$winner, loser = adv$loser, Date = adv$Date)
res <- elo_seq_bayes(standat = standat, refresh = 0, chains = 3, iter_sampling = 303)
extract_elo_b(res)
# clear expectation for c vs g
predict_new_outcome(res, "c", "g")
predict_new_outcome(res, "g", "c")
# less clear expectation for d vs f
predict_new_outcome(res, "d", "f")
# no clear expectation between e and f
predict_new_outcome(res, "f", "e")
}
