% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_seq_multigroup.R
\name{prep_seq_multigroup}
\alias{prep_seq_multigroup}
\title{prep interaction data for multi-group Stan model}
\usage{
prep_seq_multigroup(winner, loser, group, dates, presence, print_info = TRUE)
}
\arguments{
\item{winner}{character}

\item{loser}{character}

\item{group}{character}

\item{dates}{integer}

\item{presence}{character}

\item{print_info}{logical, should info messages be returned (printed)
(default is \code{TRUE}). Turn this off when running simulations.}
}
\value{
a list
}
\description{
prep interaction data for multi-group Stan model
}
\examples{
x <- group_mixture(n_ind = 20, n_mig = 4, n_grp = 3, n_int = 200)
s <- EloRating.Bayes:::prep_seq_multigroup(winner = x$x$winner,
                                           loser = x$x$loser,
                                           group = x$x$group,
                                           dates = x$x$dates,
                                           presence = x$p)
table(names(s$start_rating_index_indi))
}
