% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/presence_with_migrations.R
\name{presence_with_migrations}
\alias{presence_with_migrations}
\title{generate presence matrix with migrations}
\usage{
presence_with_migrations(
  n_ind = 10,
  n_mig = 2,
  n_grp = 2,
  n_days = 100,
  n_abs = 2
)
}
\arguments{
\item{n_ind}{integer, number of individuals}

\item{n_mig}{integer, number of migrations}

\item{n_grp}{integer, number of groups}

\item{n_days}{integer, number of days that the study covers}

\item{n_abs}{integer, number of absence-stints (see details)}
}
\value{
a list with a presence matrix (one line per day,
one column per individual), and a condensed data.frame with
the same information
}
\description{
individuals can migrate multiple times (and also return to a group)
}
\details{
\code{n_abs} determines the number of stints that are affected by absence.
An individual that migrates once has two stints. An individual that
never migrates has one stint. Any such stint can be selected for an
individual to be absent.
}
