% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principled_interactions.R
\name{principled_interactions}
\alias{principled_interactions}
\title{generate dyadic dominance interactions}
\usage{
principled_interactions(rating_grid, n_int = 26)
}
\arguments{
\item{rating_grid}{results from \code{\link{make_rating_grid}}}

\item{n_int}{numeric, number of interactions}
}
\value{
a dataframe
}
\description{
generate dyadic dominance interactions
}
\examples{
xgrid <- make_rating_grid(n_ind = 7, n_days = 100, drifts = 1)
idata <- principled_interactions(rating_grid = xgrid, n_int = 27)
head(idata)
}
