% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.bayesianelo.R
\name{summary.bayesianelo}
\alias{summary.bayesianelo}
\title{summarize bayesianelo object}
\usage{
\method{summary}{bayesianelo}(object, ...)
}
\arguments{
\item{object}{an object of class \code{"bayesianelo"},
usually the result of a call to \code{\link{elo_seq_bayes}}}

\item{...}{further arguments passed to or from other methods (ignored)}
}
\description{
summarize bayesianelo object
}
\author{
Christof Neumann
}
