% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\name{toydata}
\alias{toydata}
\alias{toyset}
\title{a toy data set}
\usage{
toydata()
}
\description{
a toy data set
}
\examples{
toyres <- toydata()

\dontrun{
set.seed(123)
x <- EloRating::randomsequence(nID = 12, avgIA = 27, presence = c(0.4, 0.4))
winner <- x$seqdat$winner
loser <- x$seqdat$loser
Date <- x$seqdat$Date
presence <- x$pres
standat <- prep_seq(winner = winner, loser = loser, Date = Date, presence = presence)
toyress <- elo_seq_bayes(standat = standat, quiet = FALSE, parallel_chains = 4, seed = 1)
}

}
