% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_elo_b.R
\name{extract_elo_b}
\alias{extract_elo_b}
\title{extract ratings}
\usage{
extract_elo_b(
  res,
  targetdate = NULL,
  make_summary = TRUE,
  quiet = TRUE,
  sel_draws = NULL,
  keep_absent = TRUE
)
}
\arguments{
\item{res}{result of \code{\link{elo_seq_bayes}}}

\item{targetdate}{vector of target dates (date format: YYYY-MM-DD). At its
default (\code{NULL}), the date of the last observed
interaction is used. If set to \code{0}, the start ratings
are returned.}

\item{make_summary}{logical, provide summary, or if \code{FALSE}: provide
posterior draws}

\item{quiet}{logical, default is \code{TRUE}: try to capture all output that
is normally printed to the console (i.e. informational messages
from cmdstanr). If \code{FALSE}: print output to console
(helpful for debugging and to monitor progress).}

\item{sel_draws}{numeric, if specified randomly select this number of
posterior draws (mostly useful for debugging and testing).
At its default (\code{NULL}) all draws are considered.
This argument is only meaningful when
\code{make_summary = FALSE}, i.e. when full posteriors
are requested.}

\item{keep_absent}{logical with default \code{TRUE}. See details section.}
}
\value{
if \code{make_summary=TRUE}: a data frame; otherwise a list of
matrices in which each matrix represent individuals in columns and
the individual draws in rows. Each list item then corresponds to
the desired date
}
\description{
extract ratings
}
\details{
\code{keep_absent} is important for how individuals are treated that were
not present (according to presence data) for a given date.
This is only relevant if
presence data was supplied to \code{\link{prep_seq}} in the first place.
If \code{keep_absent = FALSE}, absent individuals are entirely removed
from the output. If \code{keep_absent = TRUE}, these individuals will
be present in the output, but all their values will set to \code{NA}.
}
\examples{
library(EloRating)
x <- randomsequence(nID = 8, avgIA = 50)$seqdat
# make date column heterogeneous (but still ordered)
x$Date <- sort(sample(x$Date, nrow(x), TRUE))
standat <- prep_seq(winner = x$winner, loser = x$loser, Date = x$Date)
res <- elo_seq_bayes(standat, parallel_chains = 2)
extract_elo_b(res, c("2000-01-17", "2000-01-27"))
extract_elo_b(res, c("2000-01-17", "2000-01-27"), make_summary = FALSE, sel_draws = 3)

# working with presence
data(adv)
data(advpres)
sdat <- prep_seq(winner = adv$winner, loser = adv$loser, Date = adv$Date,
                 presence = advpres)
res <- elo_seq_bayes(standat = sdat, parallel_chains = 2)
# select two target dates with absent individuals
targetdate <- c("2010-01-22", "2010-02-02")
# 'c' is missing on first date, 'f' on second
extract_elo_b(res, targetdate)
# remove them from output
extract_elo_b(res, targetdate, keep_absent = FALSE)

# with full posteriors
extract_elo_b(res, targetdate, make_summary = FALSE, sel_draws = 3)
extract_elo_b(res, targetdate, make_summary = FALSE, sel_draws = 3, keep_absent = FALSE)
}
