% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_seq.R
\name{prep_seq}
\alias{prep_seq}
\alias{prep_seq_sport}
\title{prepare interaction data}
\usage{
prep_seq(
  winner,
  loser,
  Date,
  presence = NULL,
  draws = NULL,
  intensity = NULL,
  estimate_startspread = FALSE,
  estimate_k = TRUE,
  extract_dates = NULL,
  is_female = NULL
)

prep_seq_sport(
  winner,
  loser,
  Date,
  presence = NULL,
  draws = NULL,
  home_team = NULL,
  away_team = NULL,
  home_win = NULL
)
}
\arguments{
\item{winner, loser}{character vector with ids of winner and loser individuals}

\item{Date}{Date or character vector with dates (\code{"YYYY-MM-DD"})}

\item{presence}{presence (one line per interaction, which is different from
the approach in the classic EloRating package!!!)}

\item{draws}{optional logical or integer (0/1) vector with information
about draws (undecided/ties)}

\item{intensity}{optional character or factor describing type or intensity
for each interaction. Determines how many k values are estimated.
At its default (\code{NULL}), all interactions are considered of
the same type/intensity and one k value is estimated.}

\item{estimate_startspread}{logical, default is \code{FALSE}. Should the SD
of the start ratings be estimated. At its default, the SD is set to
1. When the SD is estimated, its prior is \code{exponential(1)}.}

\item{estimate_k}{logical (with default \code{TRUE}). Should \emph{k} be
estimated, or should it be set to 0 (when \code{FALSE}), which
effectively makes the system static.}

\item{extract_dates}{character or Date vector with dates. These ratings on
these dates are generated in the Stan model and therefore easier
to access during post processing. They do result in larger objects,
though. The default is \code{NULL}, i.e., store ratings after the
first interaction (i.e., not the actual start ratings) and
ratings after the last interaction ("end of the sequence").}

\item{is_female}{optional vector of 0 and 1 that encodes whether a individual
is female. If provided this vector needs to be named(!) and all
individuals in \code{winner} and \code{loser} need to be included.
Default is \code{NULL}, i.e. it is ignored.}

\item{home_team, away_team, home_win}{for sports model}
}
\value{
a list
}
\description{
then to be used with \code{\link{elo_seq_bayes}}
}
\details{
see \code{\link{elo_seq_bayes}} for examples
}
