% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_seq.R
\name{prep_seq}
\alias{prep_seq}
\title{prepare interaction data}
\usage{
prep_seq(winner, loser, Date, presence = NULL, draws = NULL)
}
\arguments{
\item{winner}{character}

\item{loser}{character}

\item{Date}{Date}

\item{presence}{presence (one line per interaction, which is different from
the approach in the classic EloRating package!!!)}

\item{draws}{optional logical or integer (0/1) vector with information
about draws (undecided/ties)}
}
\value{
a list
}
\description{
then to be used with \code{\link{elo_seq_bayes}}
}
\details{
see \code{\link{elo_seq_bayes}} for examples
}
