% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep_seq.R
\name{prep_seq}
\alias{prep_seq}
\alias{prep_seq_sport}
\title{prepare interaction data}
\usage{
prep_seq(winner, loser, Date, presence = NULL, draws = NULL)

prep_seq_sport(
  winner,
  loser,
  Date,
  presence = NULL,
  draws = NULL,
  home_team = NULL,
  away_team = NULL,
  home_win = NULL
)
}
\arguments{
\item{winner}{character}

\item{loser}{character}

\item{Date}{Date}

\item{presence}{presence (one line per interaction, which is different from
the approach in the classic EloRating package!!!)}

\item{draws}{optional logical or integer (0/1) vector with information
about draws (undecided/ties)}

\item{home_team, away_team, home_win}{for sports model}
}
\value{
a list
}
\description{
then to be used with \code{\link{elo_seq_bayes}}
}
\details{
see \code{\link{elo_seq_bayes}} for examples
}
