% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principled_interactions.R
\name{principled_interactions}
\alias{principled_interactions}
\title{generate interaction starting from actual ratings}
\usage{
principled_interactions(n_ind = 10, n_int = 100, startspread = 1)
}
\arguments{
\item{n_ind}{numeric, number of individuals}

\item{n_int}{numeric, number of interactions}

\item{startspread}{numeric, non-negative, SD of start ratings}
}
\value{
a list with two items. First a data frame with interactions and
second a vector with the start ratings.
}
\description{
for now, we assume a stable system, i.e. individuals don't change ratings
(i.e. effectively k=0)
}
\details{
if startspread = 0: internally, this will be change to 0.000001, so that no
winprob of exactly 50\% can occur. Practically, this doesn't make a
difference except for the expectation column.
}
\examples{
principled_interactions(n_ind = 6, n_int = 20, startspread = 5)
# most outcomes follow expectation:
table(principled_interactions(n_ind = 6, n_int = 1000, startspread = 5)[[1]]$expected)

# random outcomes with equal start ratings for everyone
principled_interactions(n_ind = 6, n_int = 20, startspread = 0)
table(principled_interactions(n_ind = 6, n_int = 1000, startspread = 0)[[1]]$expected)
}
