% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/principled_interactions.R
\name{rating_drift}
\alias{rating_drift}
\title{vector with rating drift (sigmoidal)}
\usage{
rating_drift(from, rat_delta, t0, t_delta = 30, k = 2, sel_id = 0)
}
\arguments{
\item{from}{numeric, rating from which the drift starts}

\item{rat_delta}{numeric, change magnitude (positive = rating increase,
negative = rating decrease)}

\item{t0}{integer, start day}

\item{t_delta}{integer (default is \code{30}), duration of the drift}

\item{k}{numeric, steepness of the slope (default is \code{2})}

\item{sel_id}{numeric, placeholder for individual (to be put in the output).
Has no effect on any calculation.}
}
\value{
a list with two items: a vector of length \code{t_delta + 1} with
new ratings, and a vector with values of arguments that were used
}
\description{
simulating 'rank changes'
}
\examples{
rating_drift(from = 2, rat_delta = 0.8, t0 = 1, t_delta = 6, k = 2)
rating_drift(from = 2, rat_delta = 0.8, t0 = 1, t_delta = 6, k = 0.1)
rating_drift(from = 2, rat_delta = 0.8, t0 = 1, t_delta = 6, k = 10)
}
