% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_group_split.R
\name{sim_group_split}
\alias{sim_group_split}
\title{simulate interactions with a group split}
\usage{
sim_group_split(
  n_ind = 15,
  n_int = 100,
  split_by_matriline = c("hard", "random")
)
}
\arguments{
\item{n_ind}{number of individuals}

\item{n_int}{number of interactions}

\item{split_by_matriline}{\code{"hard"} or \code{"random"}}
}
\value{
a list
}
\description{
group splits into two distinct subsets (groups) of individuals
}
\examples{
\dontrun{
x <- sim_group_split(n_ind = 12, n_int = 800, split_by_matriline = "hard")
standat <- prep_seq(winner = x$g1data$winner, loser = x$g1data$loser, Date = x$g1data$date)
r <- elo_seq_bayes(standat, quiet = TRUE, parallel_chains = 4, iter_sampling = 200, iter_warmup = 200)
plot_scores_longitudinal(r, plot_presence = x$p1, unc_width = 0.0)
}

}
